+++ Nos/Models/NostrConstants.swift
@@ -16,4 +16,5 @@ enum Nostr {
 
     /// See https://github.com/nostr-protocol/nips/blob/master/19.md
     static let profilePrefix = "nprofile"
+    static let eventPrefix = "nevent"
 }